Fix casing issue for tests for table finder
Sometimes, random casing will make two tables have different casing and
therefore return twice from the table finder. This is expected behaviour
as both of them need to be found and replaced case sensitively when the
tables are formatted, but the test would consider finding both versions
of the casing as an error.