Add folder for test_gcp_adapter_format_tables sql
There are two folders, where one will be formatted by a GCPAdapter that
has a gcp-project in its config and one for a GCPAdapter that does not.

The files have two queries separated by ";". The second query is the
correct formatting of the first one. If there is only one query, the
first query should remain the same after being formatted.