Remove tests that relied on old methodology
These two tests made assumptions on how the synthesization would happen,
and based the results on that. Now that the synthesization is done in a
different way, the tests fail, but they should never have been tests in
the first place. Tests that specifically check the number of values per
column are better suited to test that functionality.